Added fire charge triggered fireballs. When a player clicks a magma cream, a fireball is launched. When the fireball lands, it is nerfed so none of the surrounding environment is damaged or set on fire.
An algorithm is placed so that if the fireball snipes a player, they
immediately lose 32.5% of their health. However, if the player is < 3.2
blocks away from the explosion, they will receive some damage also; at
maximum they will lose about 22.2% of their health.

Thank you @albinoh for the great suggestion:
http://dev.bukkit.org/bukkit-plugins/king-of-the-hill/pages/wiki/arena-abilities/#c1