Players can now access a configurable GUI when trying to pick a class. If 'classes-gui' is enabled in the arena settings, and the player does not have any arguments in their command, a GUI will pop open for guidance.
Users can configure the GUI item just like any other. If more than one
item is input the config, the GUI item will revert to a grass block. The
display name of the item will reflect the class name, and will be green
if the player has permission for it or gray if the player does not have
permission. If a player clicks on a 'class' they don't have permission
for, they will be assigned a random class.