Untested Snare ability added.
The snare ability is called when a player in the arena attempts to place
a Cobweb. When an opponent steps in this cobweb, it 'bursts' and a
cobweb box 3x3x5 (LxWxH) centered around the player forms. After 2
seconds, the cobweb snare explodes with a strength almost double that of
TNT.