Untested attempt to fix common grievance of config.yml resets.
This commit offers some assistance against tabs found in config-files.
If a tab is found, generally a yaml file is reset, and all data is lost.
However, this solution throws exceptions if a tab is found ('\t' is the
definer for a tab in a Yaml file) to prevent the file from resetting.
The server admin, who checks his console when King of the Hill doesn't
load, will notice an error message given.

In the case that the config.yml file doesn't exist, a new one is
generated. This hacky solution has yet to be verified on all Operating
Systems.