Attempt to update potion handling from config files.
Previously, potions were handled like normal items, but with MC combat
update potions are now parsed with NBT tags. To accommodate, potions
need to have their duration and amplifiers stored in memory.
PotionEffects are now handled as if they are data values. The following
is the new form for PotionEffect writing.

<itemType>:<data>|<potionEffectName>@<duration>@<amplifier>(!<effect>@<duration>@<amplifier>):<amount> 

Like with normal items with durability (hello wool!), the amount is
forced for potion effects. Duration and amplifier are integer amounts.
Defaults for config files will be updated when this implementation is
fully tested.