Revert changes to Potion handling in favor of different approach to potion building. Rather than listen for custom effects, this new implementation is designed to scan for the base potion types: lingering_potion, splash_potion, or potion (unchanged) and its creative inventory potion data. That is, anything in the creative inventory menu is a valid potion.
The format for writing potions is <item>:<effect>:<amount>. To be
registered as a potion, the item must be one of the three base potion
types. The effect is analagous to the NBT tag potion names. For example,
a water bottle would be potion:water:1. An upgraded potion needs the
prefix "strong-", and an extended potion needs the prefix "long-". For
example, 3 extended invisibility potions would be
potion:long-invisibility:3.